♻️ (bc-node): remove pallet_mixnet -> move everything into one
