+++ bc-node/pallets/mixer/Cargo.toml
@@ -24,9 +24,6 @@ frame-system = { default-features = false, version = '2.0.0' }
 sp-core = { default-features = false, version = '2.0.0' }
 sp-io = { default-features = false, version = '2.0.0' }
 
-## on-chain
-pallet-mixnet = { path = '../../pallets/mixnet', default-features = false, version = '2.0.0' }
-
 # to be able to use randomness in offchain-worker
 rand_chacha = { default_features = false, version = '0.2.2' }
 rand = { default-features = false, version = "^0.7.3" }
