--- bc-node/pallets/mixnet/Cargo.toml
@@ -1,48 +1,0 @@
-[package]
-authors = ['Moritz Eck']
-description = 'Mixnet Logic for Provotum E-Voting System.'
-edition = '2018'
-homepage = 'https://github.com/meck93/provotum-mixnet'
-license = 'Unlicense'
-name = 'pallet-mixnet'
-repository = 'https://github.com/meck93/provotum-mixnet/bc-node'
-version = '2.0.0'
-
-[package.metadata.docs.rs]
-targets = ['x86_64-unknown-linux-gnu']
-
-# alias "parity-scale-code" to "codec"
-[dependencies.codec]
-default-features = false
-features = ['derive']
-package = 'parity-scale-codec'
-version = '1.3.4'
-
-[dependencies]
-frame-support = { default-features = false, version = '2.0.0' }
-frame-system = { default-features = false, version = '2.0.0' }
-
-# to be able to use on-chain randomness
-sp-core = { default-features = false, version = '2.0.0' }
-
-# to be able to use Vec
-sp-std = { default-features = false, version = '2.0.0' }
-
-# crypto library from crypto crate
-crypto = { default-features = false, path = "../../../crypto" }
-num-bigint = { default-features = false, version = "^0.3" }
-num-traits = { default-features = false, version = "^0.2" }
-
-[dev-dependencies]
-sp-io = { default-features = false, version = '2.0.0' }
-sp-runtime = { default-features = false, version = '2.0.0' }
-
-[features]
-default = ['std']
-std = [
-    'codec/std',
-    'frame-support/std',
-    'frame-system/std',
-    'sp-std/std',
-    'sp-core/std',
-]
