+++ bc-node/runtime/Cargo.toml
@@ -25,10 +25,6 @@ hex-literal = { optional = true, version = '0.3.1' }
 serde = { features = ['derive'], optional = true, version = '1.0.101' }
 
 # local dependencies
-## on-chain
-pallet-mixnet = { path = '../pallets/mixnet', default-features = false, version = '2.0.0' }
-
-## off-chain worker
 pallet-offchain-mixer = { path = '../pallets/mixer', default-features = false, version = '2.0.0' }
 
 # Substrate dependencies
@@ -82,7 +78,6 @@ std = [
     'pallet-grandpa/std',
     'pallet-randomness-collective-flip/std',
     'pallet-sudo/std',
-    'pallet-mixnet/std',
     'pallet-offchain-mixer/std',
     'pallet-timestamp/std',
     'pallet-transaction-payment/std',
